fix bug in form test
